translate role in user sidepanel
