+++ app/react/Users/components/Users.tsx
@@ -14,7 +14,20 @@ import { reset2fa } from 'app/Auth2fa/actions/actions';
 import { recoverPassword } from 'app/Auth/actions';
 import { t } from 'app/I18N';
 
-export interface UserProps {
+const roleTranslationKey = (role: string) => {
+  switch (role) {
+    case 'admin':
+      return 'Admin';
+    case 'editor':
+      return 'Editor';
+    case 'collaborator':
+      return 'Collaborator';
+    default:
+      return '';
+  }
+};
+
+interface UserProps {
   users: IImmutable<UserSchema[]>;
   userGroups: IImmutable<UserGroupSchema[]>;
   loadUsers: () => Promise<void>;
@@ -139,4 +152,6 @@ const mapDispatchToProps = {
   recoverPassword,
 };
 
+export type { UserProps };
+export { roleTranslationKey };
 export const Users = connect(mapStateToProps, mapDispatchToProps)(UsersComponent);
